tree-construction/tests22.dat

#data
<a><b><big><em><strong><div>X</a>
#errors
#document
| <html html>
|   <html head>
|   <html body>
|     <html a>
|       <html b>
|         <html big>
|           <html em>
|             <html strong>
|     <html big>
|       <html em>
|         <html strong>
|           <html div>
|             <html a>
|               "X"

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8>A</a>
#errors
#document
| <html html>
|   <html head>
|   <html body>
|     <html a>
|       <html b>
|     <html b>
|       <html div>
|         id="1"
|         <html a>
|         <html div>
|           id="2"
|           <html a>
|           <html div>
|             id="3"
|             <html a>
|             <html div>
|               id="4"
|               <html a>
|               <html div>
|                 id="5"
|                 <html a>
|                 <html div>
|                   id="6"
|                   <html a>
|                   <html div>
|                     id="7"
|                     <html a>
|                     <html div>
|                       id="8"
|                       <html a>
|                         "A"

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8><div id=9>A</a>
#errors
#document
| <html html>
|   <html head>
|   <html body>
|     <html a>
|       <html b>
|     <html b>
|       <html div>
|         id="1"
|         <html a>
|         <html div>
|           id="2"
|           <html a>
|           <html div>
|             id="3"
|             <html a>
|             <html div>
|               id="4"
|               <html a>
|               <html div>
|                 id="5"
|                 <html a>
|                 <html div>
|                   id="6"
|                   <html a>
|                   <html div>
|                     id="7"
|                     <html a>
|                     <html div>
|                       id="8"
|                       <html a>
|                         <html div>
|                           id="9"
|                           "A"

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8><div id=9><div id=10>A</a>
#errors
#document
| <html html>
|   <html head>
|   <html body>
|     <html a>
|       <html b>
|     <html b>
|       <html div>
|         id="1"
|         <html a>
|         <html div>
|           id="2"
|           <html a>
|           <html div>
|             id="3"
|             <html a>
|             <html div>
|               id="4"
|               <html a>
|               <html div>
|                 id="5"
|                 <html a>
|                 <html div>
|                   id="6"
|                   <html a>
|                   <html div>
|                     id="7"
|                     <html a>
|                     <html div>
|                       id="8"
|                       <html a>
|                         <html div>
|                           id="9"
|                           <html div>
|                             id="10"
|                             "A"

#data
<cite><b><cite><i><cite><i><cite><i><div>X</b>TEST
#errors
Line: 1 Col: 6 Unexpected start tag (cite). Expected DOCTYPE.
Line: 1 Col: 46 End tag (b) violates step 1, paragraph 3 of the adoption agency algorithm.
Line: 1 Col: 50 Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <cite>
|       <b>
|         <cite>
|           <i>
|             <cite>
|               <i>
|                 <cite>
|                   <i>
|       <i>
|         <i>
|           <div>
|             <b>
|               "X"
|             "TEST"
